[Build] Setup build conventions to use configuration caching
