+++ buildSrc/src/main/kotlin/kaff4.base.gradle.kts
@@ -1,0 +1,13 @@
+plugins {
+  id("kaff4.kotlin")
+  id("com.jaredsburrows.license")
+
+  id("kaff4.detekt")
+}
+
+licenseReport {
+  generateTextReport = true
+  generateHtmlReport = true
+  generateCsvReport = false
+  generateJsonReport = false
+}
