+++ buildSrc/src/main/kotlin/kaff4.detekt.gradle.kts
@@ -1,0 +1,40 @@
+import io.gitlab.arturbosch.detekt.Detekt
+import io.gitlab.arturbosch.detekt.DetektCreateBaselineTask
+
+plugins {
+  id("org.jetbrains.kotlin.jvm")
+  id("io.gitlab.arturbosch.detekt")
+}
+
+repositories {
+  mavenCentral()
+}
+
+val libs = extensions.getByType<VersionCatalogsExtension>().named("libs")
+
+detekt {
+  parallel = true
+  autoCorrect = true
+
+  buildUponDefaultConfig = true // preconfigure defaults
+  config.from("$rootDir/config/detekt-config.yml")
+
+  allRules = false // activate all available (even unstable) rules.
+}
+
+dependencies {
+  detektPlugins(libs.findLibrary("detekt-formatting").get())
+}
+
+tasks.withType<Detekt>().configureEach {
+  jvmTarget = "18"
+}
+tasks.withType<DetektCreateBaselineTask>().configureEach {
+  jvmTarget = "18"
+}
+
+tasks.withType<Detekt>().configureEach {
+  reports {
+    html.required.set(true)
+  }
+}
