+++ gradle.properties
@@ -1,2 +1,5 @@
 kotlin.code.style=official
 org.gradle.caching=true
+org.gradle.parallel=true
+
+org.gradle.unsafe.configuration-cache=true
