+++ kaff4-api/build.gradle.kts
@@ -1,0 +1,3 @@
+plugins {
+  id("kaff4.library")
+}
