+++ kaff4-api/kaff4-api-features/build.gradle.kts
@@ -1,3 +1,7 @@
+plugins {
+  id("kaff4.library")
+}
+
 dependencies {
   api(libs.guice)
   api(libs.misk.inject)
