+++ kaff4-cli/build.gradle.kts
@@ -2,6 +2,7 @@ import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
 
 plugins {
   application
+  id("kaff4.base")
 }
 
 dependencies {
