+++ kaff4-compression/kaff4-compression-lz4/build.gradle.kts
@@ -1,3 +1,7 @@
+plugins {
+  id("kaff4.library")
+}
+
 dependencies {
   implementation(kotlin("stdlib-jdk8"))
