+++ kaff4-core/build.gradle.kts
@@ -1,3 +1,7 @@
+plugins {
+  id("kaff4.library")
+}
+
 dependencies {
   implementation(kotlin("reflect"))
