+++ kaff4-core/kaff4-core-logging/build.gradle.kts
@@ -1,3 +1,7 @@
+plugins {
+  id("kaff4.library")
+}
+
 dependencies {
   api(libs.log4j.api)
