+++ kaff4-core/kaff4-core-model/build.gradle.kts
@@ -1,3 +1,7 @@
+plugins {
+  id("kaff4.library")
+}
+
 dependencies {
   compileOnly(kotlin("reflect"))
