+++ kaff4-core/kaff4-core-model/kaff4-core-model-api/build.gradle.kts
@@ -1,3 +1,7 @@
+plugins {
+  id("kaff4.library")
+}
+
 dependencies {
   api(libs.guava)
   api(libs.jakarta.inject.api)
