+++ kaff4-core/kaff4-core-test/build.gradle.kts
@@ -1,3 +1,7 @@
+plugins {
+  id("kaff4.library")
+}
+
 dependencies {
   implementation(kotlin("reflect"))
 
@@ -17,18 +21,4 @@ dependencies {
   implementation(project(":kaff4-core"))
   implementation(project(":kaff4-core:kaff4-core-model:kaff4-core-model-api"))
   implementation(project(":kaff4-rdf:kaff4-rdf-memory"))
-}
-
-val projectPath = project.path
-
-rootProject.allprojects {
-  val testTask by tasks.test
-
-  dependencies {
-    val thisProject = project(projectPath)
-    val testImplementationDependencies = configurations.getByName("testImplementation").dependencies
-    if (thisProject in testImplementationDependencies) {
-      testTask.jvmArgumentProviders.add { listOf("-Djunit.jupiter.extensions.autodetection.enabled=true") }
-    }
-  }
 }
