+++ kaff4-rdf/kaff4-rdf-api/build.gradle.kts
@@ -1,3 +1,7 @@
+plugins {
+  id("kaff4.library")
+}
+
 dependencies {
   api(libs.guice)
   api(libs.jakarta.inject.api)
